+++ src/Result.php
@@ -30,6 +30,13 @@ abstract class Result
     abstract protected function setData(\stdClass $data): void;
 
     /**
+     * Throws if errors are present in the result or returns an error free result.
+     *
+     * @return $this
+     */
+    abstract public function errorFree(): ErrorFreeResult;
+
+    /**
      * @return static
      */
     public static function fromResponse(Response $response): self